Refactoring: Create utils package with util modules.
