+++ src/jarvis/jarvis/utils/application_utils.py
@@ -3,14 +3,11 @@ import os
 import requests
 import traceback
 import logging
-from logging import config
-from google_speech import Speech
 from subprocess import call
-import wolframalpha
+from logging import config
 
-from jarvis.settings import GOOGLE_SPEECH, GENERAL_SETTINGS, LOG_SETTINGS
-from jarvis.settings import WOLFRAMALPHA_API
-
+from jarvis.settings import LOG_SETTINGS
+from jarvis.utils.response_utils import assistant_response
 
 jarvis_logo = "\n"\
 "      ██╗ █████╗ ██████╗ ██╗   ██╗██╗███████╗\n"\
@@ -58,31 +55,7 @@ def log(func):
     return wrapper
 
 
-def assistant_response(text):
-    """
-    Assistant response in voice or/and in text
-    :param text: string
-    """
-    if GENERAL_SETTINGS['response_in_speech']:
-        speech = Speech(text, GOOGLE_SPEECH['lang'])
-        speech.play()
-    if GENERAL_SETTINGS['response_in_text']:
-        assistant_name = GENERAL_SETTINGS['assistant_name'] + ': '
-        print('-'*48)
-        print(assistant_name + OutputStyler.CYAN + text + OutputStyler.ENDC)
-        print('-'*48 + '\n')
-
-
-def user_speech_playback(text):
-    """
-    Prints the user commands
-    """
-    print('-' * 48)
-    print('You: ' + OutputStyler.CYAN + text + OutputStyler.ENDC)
-    print('-'*48 + '\n')
-
-
-def _clear():
+def clear():
     """
     Clear stdout
     """
@@ -93,27 +66,9 @@ def start_up():
     """
     Clear the console and print the assistant logo.
     """
-    _clear()
+    clear()
     print(OutputStyler.CYAN + jarvis_logo + OutputStyler.ENDC)
     print(OutputStyler.HEADER + start_text + OutputStyler.ENDC)
-
-
-def call_wolframalpha(voice_transcript):
-    """
-    Make a request in wolframalpha API and prints the response.
-    """
-    client = wolframalpha.Client(WOLFRAMALPHA_API['key'])
-    try:
-        if WOLFRAMALPHA_API['key']:
-            res = client.query(voice_transcript)
-            assistant_response(next(res.results).text)
-            logging.debug('Successful response from Wolframalpha')
-        else:
-            assistant_response("WolframAlpha API is not working.\n"
-                               "You can get an API key from: https://developer.wolframalpha.com/portal/myapps/ ")
-    except:
-        logging.debug('There is not answer with wolframalpha')
-        assistant_response('Sorry, but I can not understand what do you want')
 
 
 def internet_connectivity_check(url='http://www.google.com/', timeout=2):
