+++ src/tests/actions_tests.py
@@ -33,7 +33,7 @@ class ActionsTests(unittest.TestCase):
         self.assertEqual(False, self.actions.wake_up_check())
 
     @patch('jarvis.action_manager.sys.exit')
-    @patch('jarvis.action_manager.SkillManager.disable_jarvis')
+    @patch('jarvis.action_manager.Skills.disable_jarvis')
     def test_execute_commands(self, mocked_exit, mocked_disable_jarvis):
 
         self.actions.latest_voice_transcript = ' '
@@ -45,7 +45,7 @@ class ActionsTests(unittest.TestCase):
         self.assertEqual(1, mocked_disable_jarvis.call_count)
 
     @patch('jarvis.actions.Actions.get_transcript')
-    @patch('jarvis.action_manager.SkillManager.enable_jarvis')
+    @patch('jarvis.action_manager.Skills.enable_jarvis')
     def test_ready_to_start(self, mocked_get_voiced_transcript, mocked_enable_jarvis):
         self.actions.latest_voice_transcript = '..'
         self.assertEqual(None, self.actions._ready_to_start())
