+++ src/tests/skills_tests.py
@@ -1,6 +1,6 @@
 import unittest
 from unittest.mock import patch
-from jarvis.skills.skill_manager import SkillManager
+from jarvis.skills.skills import Skills
 
 
 class SkillsTests(unittest.TestCase):
@@ -8,7 +8,7 @@ class SkillsTests(unittest.TestCase):
     @patch('jarvis.action_manager.wikipedia.page')
     def test_tell_me_about(self, mocked_wiki):
         words = 'tell me about google'
-        SkillManager.tell_me_about(voice_transcript=words, skill=None, tag='about')
+        Skills.tell_me_about(voice_transcript=words, skill=None, tag='about')
         self.assertEqual(1, mocked_wiki.call_count)
         words = 'about google'
-        self.assertRaises(Exception, SkillManager.tell_me_about(voice_transcript=words, skill=None, tag='about'))
+        self.assertRaises(Exception, Skills.tell_me_about(voice_transcript=words, skill=None, tag='about'))
